requirement.txt
# requirements.txt

# PDF parsing
PyMuPDF==1.23.9

# Embedding & RAG
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.4

# Voice: STT & TTS
vosk==0.3.45
sounddevice==0.4.6
pyttsx3==2.90

# GUI
PySide6==6.6.1

# General
scipy>=1.10
